machine: support samd51 boards on play.tinygo.org
