+++ src/machine/board_feather-m4.go
@@ -1,11 +1,6 @@
-// +build sam,atsamd51,feather_m4
+// +build feather_m4
 
 package machine
-
-import (
-	"device/sam"
-	"runtime/interrupt"
-)
 
 // used to reset into bootloader
 const RESET_MAGIC_VALUE = 0xf01669ef
@@ -60,37 +55,10 @@ const (
 	UART2_RX_PIN = A5
 )
 
-var (
-	UART1 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    sam.SERCOM5_USART_INT,
-		SERCOM: 5,
-	}
-
-	UART2 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    sam.SERCOM0_USART_INT,
-		SERCOM: 0,
-	}
-)
-
-func init() {
-	UART1.Interrupt = interrupt.New(sam.IRQ_SERCOM5_2, UART1.handleInterrupt)
-	UART2.Interrupt = interrupt.New(sam.IRQ_SERCOM0_2, UART2.handleInterrupt)
-}
-
 // I2C pins
 const (
 	SDA_PIN = D22 // SDA: SERCOM2/PAD[0]
 	SCL_PIN = D21 // SCL: SERCOM2/PAD[1]
-)
-
-// I2C on the Feather M4.
-var (
-	I2C0 = I2C{
-		Bus:    sam.SERCOM2_I2CM,
-		SERCOM: 2,
-	}
 )
 
 // SPI pins
@@ -98,14 +66,6 @@ const (
 	SPI0_SCK_PIN  = D25 // SCK: SERCOM1/PAD[1]
 	SPI0_MOSI_PIN = D24 // MOSI: SERCOM1/PAD[3]
 	SPI0_MISO_PIN = D23 // MISO: SERCOM1/PAD[2]
-)
-
-// SPI on the Feather M4.
-var (
-	SPI0 = SPI{
-		Bus:    sam.SERCOM1_SPIM,
-		SERCOM: 1,
-	}
 )
 
 // USB CDC identifiers
