+++ src/machine/board_pyportal.go
@@ -1,11 +1,6 @@
-// +build sam,atsamd51,pyportal
+// +build pyportal
 
 package machine
-
-import (
-	"device/sam"
-	"runtime/interrupt"
-)
 
 // used to reset into bootloader
 const RESET_MAGIC_VALUE = 0xf01669ef
@@ -111,30 +106,10 @@ const (
 	UART_RX_PIN = D0
 )
 
-var (
-	UART1 = UART{
-		Buffer: NewRingBuffer(),
-		Bus:    sam.SERCOM4_USART_INT,
-		SERCOM: 4,
-	}
-)
-
-func init() {
-	UART1.Interrupt = interrupt.New(sam.IRQ_SERCOM4_2, UART1.handleInterrupt)
-}
-
 // I2C pins
 const (
 	SDA_PIN = PB02 // SDA: SERCOM2/PAD[0]
 	SCL_PIN = PB03 // SCL: SERCOM2/PAD[1]
-)
-
-// I2C on the PyPortal.
-var (
-	I2C0 = I2C{
-		Bus:    sam.SERCOM5_I2CM,
-		SERCOM: 5,
-	}
 )
 
 // SPI pins
@@ -146,15 +121,6 @@ const (
 	NINA_MOSI = SPI0_MOSI_PIN
 	NINA_MISO = SPI0_MISO_PIN
 	NINA_SCK  = SPI0_SCK_PIN
-)
-
-// SPI on the PyPortal.
-var (
-	SPI0 = SPI{
-		Bus:    sam.SERCOM2_SPIM,
-		SERCOM: 2,
-	}
-	NINA_SPI = SPI0
 )
 
 // USB CDC identifiers
